django/requirements.txt
Django==2.2.4
djangorestframework==3.9.1
djangorestframework-gis==0.14.0
django-filter==2.2
django-url-filter==0.3.14
djangorestframework-queryfields==1.0.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.1
asn1crypto==0.24.0
cffi
cryptography==2.8
ecdsa==0.13
future==0.16.0
idna==2.6
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
python-jose==1.4.0
markdown==3.1.1
packaging==19.2
six==1.12.0
psycopg2-binary==2.7.5
python-dotenv==0.10.3
pytz==2018.5
uwsgi==2.0.18
django-allauth==0.38.0
